refactor(depsdev): split depsdev.go into per-API-flow files (struct-1)
Mechanical same-package split of depsdev.go (1567 LOC) into six per-flow
files that mirror the existing test layout:

- release.go: GetLatestReleasesForPURLs, fetchLatestRelease, fetchReleaseInfoBatch
- batch_details.go: GetDetailsForPURLs, fetchBatchPURLs, resolveRepoURLsBatch,
  repoURLFromRelatedProjects, isGemSystem, fetchPackageInfoBatch,
  countSuspiciousMavenPURLs, fetchProjectsBatch, buildFinalResults,
  buildBasicResult, buildCompleteResult, convertRepoURLToProjectKey,
  normalizeRepoURLForProject
- package_info.go: fetchPackageInfo, fetchPURLRaw, tryMavenSearchFallback,
  stripTrailingVersion
- dependents.go: FetchDependentCount, FetchDependentCountBatch
- licenses.go: GetPackageVersionLicenses, collectVersionLicenses
- advisory.go: FetchAdvisory, FetchAdvisoriesBatch

depsdev.go retains: errPURLNotFound, DepsDevClient, NewDepsDevClient,
With* options, purlpkgToParsed, ExtractRepositoryURLFromLinks, truncateString.

No signature changes. go test ./internal/infrastructure/depsdev/... passes
unchanged.
